IN_DEVELOPMENT - bug 241: As an administrative user, I would like to launch jobs from the portal http://129.67.24.160/bugzilla/show_bug.cgi?id=241
Also included toggle to comment out web map for now as metacata / EDIT map servers are slowing down the tests too much

+++ emonocot-model/src/main/java/org/emonocot/model/job/Job.java
@@ -67,7 +67,7 @@ public class Job extends Base {
    /**
     *
     */
-   private Integer jobId;
+   private Long jobId;
    /**
     *
     */
@@ -215,14 +215,14 @@ public class Job extends Base {
     /**
      * @return the jobId
      */
-    public Integer getJobId() {
+    public Long getJobId() {
         return jobId;
     }
 
     /**
      * @param newJobId the jobId to set
      */
-    public void setJobId(Integer newJobId) {
+    public void setJobId(Long newJobId) {
         this.jobId = newJobId;
     }
